Remove BCMath prereq.
We first check if we can decode the "big" uint with normal PHP
integers. If not, we check for GMP or BCMath. If those are not
available, we throw an exception.

Most databases will not have such large numbers, so it doesn't
make sense to have BCMath or GMP as a hard requirement.