autoload.php: new PSR-4 autoloader for the MaxMind\Db namespace.
Since we follow PSR-4, it is easy to deduce the PHP file that defines
a class from that class's name. If used to install the library,
composer will produce an "autoload.php" file for you that does just
that. However, when composer is not used, the end user is left with no
way to load the class files that the library requires.

This commit adds a PSR-4 autoload.php file that will work without
composer. To use it, simply require("/path/to/autoload.php") in your
project. This has several benefits:

  1. The release tarball of the library is usable out-of-the-box,
     which means that end-users can verify its hash to ensure that
     the code they're running came from where they think it did.

  2. The test suite doesn't require internet access to run.

  3. As a result of the first two items, the library can be installed
     using other (system-level, distribution) package managers.

Closes: https://github.com/maxmind/MaxMind-DB-Reader-php/issues/56