tests/bootstrap.php: use our autoload.php instead of composer's.
The "autoload.php" file that we ship is always present, and doesn't
require composer (or internet access) to create. That makes it a
better choice to use in the test suite; for example, the tests can be
run directly from a git clone. This commit tells the tests to use our
custom autoloader and not the one that composer would create.